Add API URL routing and home view implementation
- Created a new `api_urls.py` file to define API routes, including paths for the todo app and schema documentation.
- Moved the home view function to a new `views.py` file, maintaining its functionality and context data for rendering.
- Updated the main `urls.py` to include the new API routing structure.

These changes improve the organization of the codebase by separating API routing from the main URL configuration and enhancing the clarity of the home view implementation.
